feat: centralize wp-post storage artifacts
