+++ packages/wp-json-ast/src/resource/index.ts
@@ -4,9 +4,7 @@ export * from './common/string';
 export * from './errors';
 export * from './cache';
 export * from './query';
-export * from './wp-post/query';
-export * from './wp-post/mutation';
-export * from './wp-post/routes';
+export * from './wp-post';
 export * from './storage/wp-option';
 export * from './storage/transient';
 export * from './wp-taxonomy';
