+++ packages/wp-json-ast/src/resource/wp-post/index.ts
@@ -1,0 +1,8 @@
+export * from './query';
+export * from './mutation';
+export * from './routes';
+export { buildWpPostStorageArtifacts } from './buildWpPostStorageArtifacts';
+export type {
+	BuildWpPostStorageArtifactsOptions,
+	WpPostStorageArtifacts,
+} from './buildWpPostStorageArtifacts';
